+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next
 
+### Calendar
+
+- Added `disableYearPagination` prop to disable year pagination arrows.
+
 ### DateRangeDualTextInput
 - Add `calendarProps` which will be passed along to the internal Calendar-component, and can be used to specify enabled dates and `onDayClick` callback.